requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
Babel==2.5.0
beautifulsoup4==4.4.0
billiard==3.3.0.23
bleach==1.5.0
celery==3.1.25
click==6.7
coverage==4.4.1
defusedxml==0.5.0
diff-match-patch==20121119
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.10
django-annoying==0.10.3
django-autoslug==1.9.3
django-bootstrap-form==3.2.1
django-bootstrap3==9.0.0
django-bower==5.2.0
django-celery==3.2.1
django-cors-headers==1.1.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.5
django-extra-views==0.9.0
django-filer==1.2.7
django-filter==1.0.2
django-formtools==2.0
django-geojson==2.10.0
django-geoposition==0.3.0
django-google-maps==0.2.1
django-hero-slider==1.3.1
django-hvad==1.8.0
django-import-export==0.5.1
django-libs==1.67.12
django-mptt==0.8.7
django-phonenumber-field==1.3.0
django-polymorphic==1.0.2
django-scheduler==0.8.3
django-sendsms==0.2.3
django-sql-explorer==1.1.1
django-tables2==1.10.0
django-taggit==0.22.1
django-twilio==0.8.0
djangorestframework==3.6.3
easy-thumbnails==2.3
et-xmlfile==1.0.1
Flask==0.10.1
gunicorn==19.6.0
heroku==0.1.4
html5lib==0.9999999
icalendar==3.11.4
itsdangerous==0.24
jdcal==1.3
Jinja2==2.9.6
jsonfield==1.0.3
kombu==3.0.37
Markdown==2.6.7
MarkupSafe==1.0
mock==2.0.0
numpy==1.13.1
oauthlib==2.0.2
odfpy==1.3.5
olefile==0.44
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pbr==3.0.0
phonenumberslite==8.8.0
Pillow==4.0.0
psycopg2==2.6.2
py==1.4.33
PyJWT==1.5.0
pyparsing==2.2.0
PySocks==1.6.7
python-dateutil==2.6.0
python-decouple==2.3
python-firebase==1.2
python-http-client==3.0.0
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
redis==2.10.3
reportlab==3.4.0
requests==2.14.2
requests-oauthlib==0.8.0
sendgrid==5.0.0
six==1.10.0
social-auth-app-django==1.1.0
social-auth-core==1.3.0
sorl-thumbnail==12.4a1
South==1.0
sqlparse==0.2.3
static3==0.7.0
tablib==0.11.2
tox==1.7.2
twilio==6.0.0
Twisted==15.3.0
unicodecsv==0.14.1
Unidecode==0.4.20
Unipath==1.1
urllib3==1.21.1
vine==1.1.4
virtualenv==15.1.0
vobject==0.9.5
webencodings==0.5.1
Werkzeug==0.12.2
whitenoise==3.2.2
WTForms==2.1
xlrd==1.1.0
xlwt==1.3.0
zope.interface==4.4.0
